Update of code
Some tests and workarounds were make in order to see how the characterImage was loaded in the cells of the collectionView and the tableView. At the end, the "new" code was added but commented just to let know that those options exists and the behavior/performance of the app is still the same than version 1.